+++ nifty/src/main/java/de/lessvoid/nifty/internal/canvas/CommandLineTo.java
@@ -3,10 +3,10 @@ package de.lessvoid.nifty.internal.canvas;
 import de.lessvoid.nifty.internal.render.batch.BatchManager;
 
 public class CommandLineTo implements Command {
-  private final float x;
-  private final float y;
+  private final double x;
+  private final double y;
 
-  public CommandLineTo(final float x, final float y) {
+  public CommandLineTo(final double x, final double y) {
     this.x = x;
     this.y = y;
   }
